Added new unittests for cache.validate
